% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_dependencies.R
\name{find_r_files}
\alias{find_r_files}
\title{Get all R (.R or .Rmd) files in directory}
\usage{
find_r_files(directory = ".")
}
\arguments{
\item{directory}{directory to search in. Defaults to "."}
}
\value{
vector of all the file paths (relative to directory)
}
\description{
Get all R (.R or .Rmd) files in directory
}
\examples{
find_r_files()
}
